fug.jl

fug.geoMean — Function

geoMean(kLKd::Float64, kLKb::Float64, kHKd::Float64, kHKb::Float64)

returns the geometric mean of k values

\[\sqrt{\frac{LK_d \cdot LK_b}{HK_d \cdot HK_b}}\]

Arguments

'kLKd': k value of light key in distilate

'kLKb': k value of light key in bottoms

'kHKd': k value of heavy key in distilate

'kHKb': k value of heavy key in bottoms

fug.fenske — Function

fenske(xLKd, kLKd, xHKd, kHKd, xLKb, kLKb, xHKb, kHKb)

returns the theoretical minimun number of trays for the distillation column

\[n = \frac{log_{10}(\frac{xLKd \cdot xHKb}{xLKb \cdot xHKd})}{log_{10}(\alpha_{avg})}\]

Arguments

'xLKd': x value of light key in distilate

'kLKd': k value of light key in distilate

'xHKd': x value of heavy key in distilate

'kHKd': k value of heavy key in distilate

'xLKb': x value of light key in bottoms

'kLKb': k value of light key in bottoms

'xHKb': x value of heavy key in bottoms

'kHKb': k value of heavy key in bottoms

fug.nonKeyComp — Function

nonKeyComp(alpha, nmin, F, D, B, xHKd, xHKb, xfi)

returns the molar fraction of a non-key component in both the bottoms and the distillate

(xBi, xDi)

by solving the 2 equations at the same time:

\[\frac{x_{di}}{x_{bi}} = \alpha^{n_{min}} \cdot \frac{x_{HK,d}}{x_{HK,b}}\]

\[F_i = D_i + B_i\]

done by using the following optimization:

\[Min[(\frac{\alpha^{N_{min}} \cdot x_{HK,d}}{x_{HK,b}} - \frac{x_{i,d} \cdot B}{F \cdot x_{i,f} - D \cdot x_{i,d}})^2]\]

Arguments

'alpha': relative volitility of component with the heavy key

'nmin': the theoretical minimum number of stages for the column, get from Fenske equation

'F': feed molar flow rate

'D': distillate molar flow rate

'B': bottoms molar flow rate

'xHKd': x value of the heavy key in the distillate

'xHKb': x value of the heavy key in the bottoms

'xfi': the x value of the non-key component in the feed

fug.underwood — Function

underwood(kDist, kBot, kHKd, kHKb, xFeed, qual, xDist)

The Underwood equation for finding the minimum reflux ratio by solving the first equation subject to the second condition:

\[\sum_{i=1}^n \frac{\alpha_i \cdot x_{i,f}}{\alpha_i - \Theta} = 1-q\]

\[R_{min} = \sum_{i=1}^n \frac{\alpha_i \cdot x_{i,d}}{\alpha_i - \Theta} - 1\]

Arguments

'kDist': vector of k values for all components in the distillate

'kBot': vector of k values for all components in the bottoms

'kHKd': k value of heavy key in distillate

'kHKb': k value of heavy key in bottoms

'xFeed': vector of x values for the feed

'qual': quality of the feed stream (percentage of vapor if in VLE)

'xDist': vector of x values for the distillate

fug.gilliand — Function

gilliand(ract, rmin, nmin)

The Gilliand correlation for the actual number of stages. Solve the following equation for N

\[\frac{N-N_{min}}{N+1} = 0.75 \cdot [1-(\frac{R-R_{min}}{R+1})^{0.566}]\]

Arguments

'ract': The actual reflux ratio, typically 1.1-1.4 times the minimum reflux ratio

'rmin': Minimum reflux ratio (Underwood equation)

'nmin': Theoretical minimum number of stages (Fenske equation)

fug.kirkbride — Function

kirkbride(B,D,xHKf,xLKf,xHKd,xLKb,nTot)

The Kirkbride function for finding the optimial feed stage. Solves the first equation subject to the constraint:

\[ln(\frac{N_d}{N_b}) = 0.206 \cdot ln(\frac{B \cdot x_{HK,f} \cdot x_{LK,b}^2}{D \cdot x_{LK,f} \cdot x_{HK,d}^2})\]

\[N_{tot} = N_d + N_b\]

Arguments

'B': bottoms molar flow rate

'D': distillate molar flow rate

'xHKf': x value of heavy key in feed

'xLKf': x value of light key in feed

'xHKd': x value of heavy key in distillate

'xLKb': x value of light key in bottoms

'nTot': total number of stages in column

fug.diameter — Function

diameter(Fha, Ff, Fst, Cst, denl, deng, f, Ad, At, G)

function for finding the diameter of the column based on fluid properities. Solves equation for:

\[D_T = \sqrt{\frac{4G}{fU_f \pi (1-(A_d/A_T)) \rho_g}}\]

\[U_f = C \cdot \sqrt{\frac{\rho_l-\rho_g}{\rho_g}}\]

\[C = C_{st} \cdot F_{st} \cdot F_f \cdot F_{ha} \]

Arguments

'F_ha': Hole factor, 1 for valve and bubble cap trays;

fug.fugk — Function

fugk(comp, xFeed, kBot, kDist, HKi, LKi, splitFracHK, splitFracLK, F, B, D, quality, RRHeuristic)

main function for the Fenske-Underwood-Gilliand-Kirkbride function. Please make sure to put your components in the correct order.

Arguments

'comp': vector of component names in order from lowest to highest volitility

'xFeed': vector of fractions for each component in comp

'kBot': vector of k values for the bottoms of the column from lowest to highest

'kDist': vector of k values for the distillate of the column from lowest to highest

'HKi': Index of the heavy key (ie index in the comp list)

'LKi': Index of the light key (ie index in the comp list)

'splitFracHK': The fraction of the heavy key that will go to the bottoms. If 90\% of the heavy key goes to the bottoms then 0.9

'splitFracLK': The fraction of the light key that will go to the distilate

'F': Feed molar feed rate

'B': Bottoms molar flow rate

'D': Distilate molar flow rate

'quality': vapor fraction of the feed

'RRHeuristic': reflux ratio heuristic value, typically in range [1.2,1.4]